added magento base url

+++ bin/magento2-start.sh
@@ -2,16 +2,27 @@
 echo "Initializing setup..."
 cd /var/www/html
 if [ -f ./app/etc/config.php ] || [ -f ./app/etc/env.php ]; then
+    echo "Update Magento 2 base url to $MAGENTO_BASE_URL"
+    /usr/bin/php ./bin/magento setup:store-config:set --base-url="$MAGENTO_BASE_URL"
+    /usr/bin/php ./bin/magento setup:store-config:set --base-url-secure="$MAGENTO_BASE_URL"
+    /usr/bin/php ./bin/magento cache:flush
     echo "It appears Magento is already installed (app/etc/config.php or app/etc/env.php exist). Exiting setup..."
     exit
 fi
 
 echo "Download Magento2 ..."
-curl -L http://pubfiles.nexcess.net/magento/ce-packages/magento2-2.1.2.tar.gz | tar xzf - -o -C .
-#/usr/bin/php /usr/local/bin/composer create-project --repository-url=https://repo.magento.com/ magento/project-community-edition=2.1.2 .
+if [ -f ./magento2-2.1.2.tar.gz ]; then
+    tar xzf ./magento2-2.1.2.tar.gz
+else
+    curl -L http://pubfiles.nexcess.net/magento/ce-packages/magento2-2.1.2.tar.gz | tar xzf - -o -C .
+fi
+#curl -L http://pubfiles.nexcess.net/magento/ce-packages/magento2-2.1.2.tar.gz | tar xzf - -o -C .
 echo "Download Magento2 Complete"
 
+chmod -R 777 ./var
+chown -R www-data. /var/www/html
 chmod +x ./bin/magento
+
 echo "Installing composer dependencies..."
 /usr/bin/php /usr/local/bin/composer update
 
@@ -24,11 +35,12 @@ mysql -e "FLUSH PRIVILEGES;"
 
 echo "Running Magento 2 setup script..."
 /usr/bin/php ./bin/magento setup:install \
+  --backend-frontname=adminlogin \
   --db-host=localhost \
   --db-name=magento2 \
   --db-user=magento2 \
   --db-password=magento2 \
-  --base-url=http://localhost/ \
+  --base-url=$MAGENTO_BASE_URL \
   --admin-firstname=Admin \
   --admin-lastname=Admin \
   --admin-email=admin@admin.com \
